IDao now uses long for ID, implemented Update and ListAll

+++ Wardrobe-Program/Application/ChangeNameController.cs
@@ -13,7 +13,7 @@ namespace Wardrobe_Program
         //Assumes that an id is given as first parameter, and name is given as second parameter, i.e. "change-name 44 A-new-name"
         public void Handle(Command command) {
             UserInterface.Instance.Print("This should change the garment name");
-            int id = Convert.ToInt32(command.Parameters[0]);
+            long id = Convert.ToInt64(command.Parameters[0]);
             Garment garmentToChange = _garmentDao.Retrieve(id);
             garmentToChange.Name = command.Parameters[1];
             UserInterface.Instance.Print($"Garment name is now: {garmentToChange.Name}");
