+++ Wardrobe-Program/Application/ChangePriceController.cs
@@ -15,9 +15,9 @@ namespace Wardrobe_Program
         public void Handle(Command command)
         {
             UserInterface.Instance.Print("This should change the garment price");
-            int id = Convert.ToInt32(command.Parameters[0]);
+            long id = Convert.ToInt64(command.Parameters[0]);
             Garment garmentToChange = _garmentDao.Retrieve(id);
-            garmentToChange.Price = Convert.ToInt32(command.Parameters[1]);
+            garmentToChange.Price = Convert.ToSingle(command.Parameters[1]);
             UserInterface.Instance.Print($"Garment price is now: {garmentToChange.Price}");
         }
     }
