+++ Wardrobe-Program/Application/ChangeSizeController.cs
@@ -17,7 +17,7 @@ namespace Wardrobe_Program
 		public void Handle(Command command)
 		{
 			UserInterface.Instance.Print("This should change the garment size");
-			int id = Convert.ToInt32(command.Parameters[0]);
+			long id = Convert.ToInt64(command.Parameters[0]);
 			Garment garmentToChange = _garmentDao.Retrieve((id));
 			garmentToChange.Size = command.Parameters[1];
 			UserInterface.Instance.Print($"Garment size is now: {garmentToChange.Size}");
